Clean up SourcifyEventManager
