+++ services/server/src/server/middlewares/GenericErrorHandler.ts
@@ -1,6 +1,5 @@
 import * as HttpStatus from "http-status-codes";
 import { Request, Response } from "express";
-import { SourcifyEventManager } from "../../common/SourcifyEventManager/SourcifyEventManager";
 
 export default function genericErrorHandler(
   err: any,
